requirements.txt
asgiref==3.7.2
autopep8==2.0.2
boto3==1.28.7
botocore==1.31.7
Django==4.2.3
django-storages==1.13.2
django-tinymce==3.6.1
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
whitenoise==6.5.0
